MWPW-189894: relocate some of functions.
